requirements.txt
bidict==0.23.1
blinker==1.9.0
boto3==1.38.6
botocore==1.38.6
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.5.1
greenlet==3.2.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.2.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
s3transfer==0.12.0
simple-websocket==1.1.0
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
